scripts/temp_migration.py
import os
from supabase import create_client
from dotenv import load_dotenv

load_dotenv()

url = os.environ.get("SUPABASE_URL")
key = os.environ.get("SUPABASE_SERVICE_ROLE_KEY")
supabase = create_client(url, key)

sql = """
ALTER TABLE public.bin DROP COLUMN IF EXISTS bin_date CASCADE;
ALTER TABLE public.egg ADD COLUMN IF NOT EXISTS intake_date DATE NOT NULL DEFAULT CURRENT_DATE;
"""

# Supabase python client doesn't have a direct 'execute raw sql' easily for DDL in the standard client.
# It usually goes through RPC or the dashboard. 
# However, I can try to use the 'rpc' method if a 'run_sql' function exists in their setup, 
# but usually it doesn't.
# I will assume the user will apply the master schema or I'll just update the app code 
# to be compatible with both for a moment, or ask.
# Actually, I'll just update the .py files.

print("Schema change script started (Simulation).")
