requirements.txt
# =============================================================================
# requirements.txt: Stable Production Dependencies
# Project:     Incubator Vault v8.1.1 — WINC (Clinical Sovereignty Edition)
# Description: Fixed version constraints to prevent breaking changes in the cloud.
# =============================================================================

streamlit>=1.35.0
supabase>=2.4.0
python-dotenv>=1.0.1
pandas>=2.1.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
httpx>=0.25.0

# Testing & Development
pytest>=9.0.0
pytest-mock>=3.14.0
pytest-watcher>=0.4.0
